requirements.txt
attrs==21.4.0
Automat==20.2.0
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
constantly==15.1.0
crypto==1.4.1
cssselect2==0.6.0
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.0.2
fonttools==4.33.3
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
Naked==0.1.31
numpy==1.22.2
opencv-python==4.5.5.62
openpyxl==3.0.10
pandas==1.4.3
pdfkit==1.0.0
pi==0.1.2
Pillow==9.1.1
pipreqs==0.4.11
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
sllurp==0.5.1
smokesignal==0.7.0
tinycss2==1.1.1
tornado==6.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
urllib3==1.26.9
weasyprint==55.0
webencodings==0.5.1
Werkzeug==2.0.2
wkhtmltopdf==0.2
yarg==0.1.9
zope.interface==5.4.0
zopfli==0.2.1
